app/src/test/java/org/openlmis/core/view/holder/ArchivedDrugsViewHolderTest.java
package org.openlmis.core.view.holder;

import android.view.LayoutInflater;
import android.view.View;

import org.junit.Before;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.openlmis.core.LMISTestApp;
import org.openlmis.core.LMISTestRunner;
import org.openlmis.core.R;
import org.openlmis.core.model.Product;
import org.openlmis.core.model.builder.ProductBuilder;
import org.openlmis.core.view.viewmodel.StockCardViewModel;
import org.openlmis.core.view.viewmodel.StockCardViewModelBuilder;
import org.robolectric.RuntimeEnvironment;

import static org.assertj.core.api.Assertions.assertThat;

@RunWith(LMISTestRunner.class)
public class ArchivedDrugsViewHolderTest {

    private ArchivedDrugsViewHolder viewHolder;
    private String queryKeyWord = null;
    private StockCardViewModel viewModel;

    @Before
    public void setUp() {
        View itemView = LayoutInflater.from(RuntimeEnvironment.application).inflate(R.layout.item_archive_drugs, null, false);
        viewHolder = new ArchivedDrugsViewHolder(itemView);

        Product product = new ProductBuilder().setPrimaryName("Lamivudina 150mg").setCode("08S40").setStrength("10mg").setType("VIA").build();
        viewModel = new StockCardViewModelBuilder(product)
                .setQuantity("10")
                .setChecked(false)
                .setType("Embalagem")
                .setSOH(123L)
                .build();
    }

    @Test
    public void shouldShowProductNameAndStyledUnit() {
        ((LMISTestApp) RuntimeEnvironment.application).setFeatureToggle(true);

        viewHolder.populate(viewModel, queryKeyWord);

        assertThat(viewHolder.tvProductName.getText().toString()).isEqualTo("Lamivudina 150mg [08S40]");
        assertThat(viewHolder.tvProductUnit.getText().toString()).isEqualTo("10mg VIA");
    }
}